Refactor dashboard data flow & UI components
Consolidate dashboard data and modernize UI wiring: introduce getDashboard() to fetch a single DashboardData payload (with a temporary awards override until backend ships), and thread that data via DashboardPage -> DashboardView -> Hero/PickStatus/Leaderboard. Add DashboardLoading and wrap content in Suspense for client-side loading, and move PickStatus and Leaderboard into client components (removing legacy CombinedLeaderboardCard and PickStatusCard). Share and simplify animations (card fade-up and tutorial background loops), unify countdown badge, add CSS var for tutorial scroll offset, update hero assets, and clean up unused helpers/types (remove mock and an unused countdown formatter). Also export DashboardLoading and update type exports accordingly.
