+++ src/app/page.tsx
@@ -1,7 +1,19 @@
-import { DashboardView } from "@/features/dashboard";
+import { Suspense } from "react";
+
+import { DashboardLoading, DashboardView } from "@/features/dashboard";
+import { getDashboard } from "@/features/dashboard/api/dashboard.api";
 import { getCurrentUser } from "@/lib/auth";
 
-export default async function DashboardPage() {
+async function DashboardContent() {
   const user = await getCurrentUser();
-  return <DashboardView isLoggedIn={!!user} />;
+  const data = await getDashboard(!!user);
+  return <DashboardView isLoggedIn={!!user} data={data} currentUserId={user?.id ?? null} />;
+}
+
+export default function DashboardPage() {
+  return (
+    <Suspense fallback={<DashboardLoading />}>
+      <DashboardContent />
+    </Suspense>
+  );
 }
