+++ src/features/dashboard/animations/card.animations.ts
@@ -10,26 +10,8 @@ type CardAnimationProps = {
   card: gsap.DOMTarget;
 };
 
-export function pickStatusCardAnimation({ card }: CardAnimationProps) {
-  const tl = gsap.timeline({
-    defaults: {
-      ease: DEFAULT_EASE,
-    },
-  });
-
-  tl.fromTo(card, fadeUp, {
-    opacity: 1,
-    y: 0,
-    duration: DEFAULT_DURATION * 1.5,
-  });
-
-  return () => {
-    tl.kill();
-    tl.scrollTrigger?.kill();
-  };
-}
-
-export function leaderboardCardAnimation({ card }: CardAnimationProps) {
+// Shared fade-up reveal for dashboard cards (pick status, leaderboard).
+export function cardFadeUpAnimation({ card }: CardAnimationProps) {
   const tl = gsap.timeline({
     defaults: {
       ease: DEFAULT_EASE,
