+++ src/features/dashboard/components/AuthHero.tsx
@@ -6,18 +6,22 @@ import { useTranslations } from "next-intl";
 import { Button } from "@/shared/components/ui/button";
 
 import { useCountdown } from "../hooks/useCountdown";
+import { isAllPickemComplete } from "../lib/pickStatusDerivations";
 import { TOURNAMENT_START_DATE } from "../lib/tournamentConfig";
-import type { DashboardStats } from "../types/dashboard.types";
+import type { DashboardStats, Match, PickemProgress, Team } from "../types/dashboard.types";
 
+import { HeroCountdownBadge } from "./GuestHeroCountdown";
 import { HeroCard } from "./HeroCard";
 import { UserStatsRow } from "./UserStatsRow";
 
 type Props = {
+  pickedChampion: Team | null;
   stats: DashboardStats | null;
-  isPickemComplete: boolean;
+  nextMatch: Match | null;
+  pickemProgress: PickemProgress | null;
 };
 
-export function AuthHero({ stats, isPickemComplete }: Props) {
+export function AuthHero({ pickedChampion, stats, nextMatch, pickemProgress }: Props) {
   const t = useTranslations("dashboard.hero");
   const countdown = useCountdown(TOURNAMENT_START_DATE);
 
@@ -25,25 +29,21 @@ export function AuthHero({ stats, isPickemComplete }: Props) {
     document.querySelector("#tutorial")?.scrollIntoView({ behavior: "smooth", block: "start" });
   };
 
-  const badge = (
-    <span className="w-fit p-1.5 sm:p-3 rounded-lg bg-muted text-xs sm:text-sm text-muted-foreground">
-      {!countdown.isExpired ? t("countdown", { days: countdown.days, hours: countdown.hours, minutes: countdown.minutes }) : t("badge")}
-    </span>
-  );
-
+  const isPickemComplete = pickemProgress ? isAllPickemComplete(pickemProgress) : false;
   const primaryCtaLabel = isPickemComplete || countdown.isExpired ? t("cta.seeMyBrackets") : t("cta.finishBracket");
 
-  const bottomContent = stats ? (
-    <UserStatsRow stats={stats} />
-  ) : (
-    <div className="flex flex-col gap-2 py-2 sm:py-3">
-      <p className="text-sm text-foreground">{t("welcome.subtitle")}</p>
-      <Button onClick={scrollToTutorial} className="w-full md:w-fit">
-        {t("welcome.cta")}
-        <ArrowRight className="size-4" />
-      </Button>
-    </div>
-  );
+  const bottomContent =
+    stats && isPickemComplete ? (
+      <UserStatsRow pickedChampion={pickedChampion} stats={stats} nextMatch={nextMatch} />
+    ) : (
+      <div className="flex flex-col gap-2 py-2 sm:py-3">
+        <p className="text-sm text-foreground">{t("welcome.subtitle")}</p>
+        <Button onClick={scrollToTutorial} className="w-full md:w-fit">
+          {t("welcome.cta")}
+          <ArrowRight className="size-4" />
+        </Button>
+      </div>
+    );
 
-  return <HeroCard badge={badge} primaryCta={{ href: "/bracket", label: primaryCtaLabel }} bottomContent={bottomContent} />;
+  return <HeroCard badge={<HeroCountdownBadge />} primaryCta={{ href: "/bracket", label: primaryCtaLabel }} bottomContent={bottomContent} />;
 }
