+++ src/features/dashboard/components/DashboardView.tsx
@@ -1,6 +1,7 @@
 import { FloatingShapes } from "@/shared/components/FloatingShapes";
 
 import { floatingShapes } from "../lib/dashboardFloatingShapes";
+import type { DashboardData } from "../types/dashboard.types";
 
 import { HeroSection } from "./HeroSection";
 import { LeaderboardSection } from "./LeaderboardSection";
@@ -9,22 +10,30 @@ import { TutorialSection } from "./TutorialSection";
 
 type Props = {
   isLoggedIn: boolean;
+  data: DashboardData | null;
+  currentUserId: string | null;
 };
 
-export function DashboardView({ isLoggedIn }: Props) {
+export function DashboardView({ isLoggedIn, data, currentUserId }: Props) {
   return (
     <div className="relative flex flex-col">
-      <HeroSection isLoggedIn={isLoggedIn} />
+      <HeroSection
+        isLoggedIn={isLoggedIn}
+        pickedChampion={data?.picked_champion ?? null}
+        stats={data?.stats ?? null}
+        nextMatch={data?.next_match ?? null}
+        pickemProgress={data?.progress.pickem ?? null}
+      />
 
       <section className="border-t border-border">
         <div className="container mx-auto px-4 py-6 sm:px-6 lg:px-8">
           <FloatingShapes shapes={floatingShapes} />
           <div className="flex flex-col lg:flex-row gap-6">
             <div className="flex-1">
-              <PickStatusSection isLoggedIn={isLoggedIn} />
+              <PickStatusSection isLoggedIn={isLoggedIn} progress={data?.progress ?? null} />
             </div>
             <div className="flex-1">
-              <LeaderboardSection isLoggedIn={isLoggedIn} />
+              <LeaderboardSection leaderboard={data?.leaderboard ?? null} currentUserId={currentUserId} />
             </div>
           </div>
         </div>
