+++ src/features/dashboard/components/UserStatsRow.tsx
@@ -7,10 +7,12 @@ import { useLocale, useTranslations } from "next-intl";
 
 import { formatShortDate } from "@/shared/lib/dates";
 
-import type { DashboardStats } from "../types/dashboard.types";
+import type { DashboardStats, Match, Team } from "../types/dashboard.types";
 
 type Props = {
+  pickedChampion: Team | null;
   stats: DashboardStats;
+  nextMatch: Match | null;
 };
 
 type StatItemProps = {
@@ -28,7 +30,7 @@ function StatItem({ label, icon, iconContent, value }: StatItemProps) {
       </div>
 
       <div className="flex flex-col sm:text-center xl:text-start min-w-0 flex-1 gap-1">
-        <span className="text-xs sm:text-[11px] xl:text-xs h-5 md:h-auto flex items-center uppercase tracking-wider text-muted-foreground leading-none md:justify-center xl:justify-start">
+        <span className="text-xs sm:text-2xs xl:text-xs h-5 md:h-auto flex items-center uppercase tracking-wider text-muted-foreground leading-none md:justify-center xl:justify-start">
           {label}
         </span>
         {value}
@@ -52,16 +54,14 @@ function CompetitionStat({ label, rank, points, icon }: { label: string; rank: n
   );
 }
 
-export function UserStatsRow({ stats }: Props) {
+export function UserStatsRow({ pickedChampion, stats, nextMatch }: Props) {
   const t = useTranslations("dashboard.stats");
   const locale = useLocale();
 
-  const championName = stats.picked_champion?.name[locale] ?? stats.picked_champion?.name["en"] ?? "—";
-
-  const nextMatch = stats.next_match;
+  const championName = pickedChampion?.name[locale] ?? pickedChampion?.name["en"] ?? "—";
   const kickoffDate = nextMatch ? formatShortDate(new Date(nextMatch.kickoff_at), locale) : null;
-  const homeFifaCode = nextMatch?.home_team?.fifa_code ?? "";
-  const awayFifaCode = nextMatch?.away_team?.fifa_code ?? "";
+  const homeTeam = nextMatch?.teams.home ?? null;
+  const awayTeam = nextMatch?.teams.away ?? null;
 
   return (
     <>
@@ -69,9 +69,7 @@ export function UserStatsRow({ stats }: Props) {
       <StatItem
         label={t("predictedChampion")}
         icon={Trophy}
-        iconContent={
-          stats.picked_champion?.flag_url && <img src={stats.picked_champion.flag_url} alt={stats.picked_champion.fifa_code} className="size-7 object-contain" />
-        }
+        iconContent={pickedChampion?.flag_url && <img src={pickedChampion.flag_url} alt={pickedChampion.fifa_code} className="size-7 object-contain" />}
         value={<span className="font-heading text-sm md:text-base xl:text-lg font-bold leading-tight truncate">{championName}</span>}
       />
 
@@ -81,17 +79,17 @@ export function UserStatsRow({ stats }: Props) {
       {/* 3 — Match picks: rank + points */}
       <CompetitionStat icon={Target} label={t("matchRank")} rank={stats.match.rank} points={stats.match.points} />
 
-      {/* 4 — Next match: teams as main text, date as muted */}
+      {/* 4 — Next match */}
       {nextMatch && kickoffDate && (
         <StatItem
           label={`${t("nextMatch")} · ${kickoffDate}`}
           icon={CalendarDays}
           value={
-            <div className="flex items-center sm:justify-center xl:justify-start gap-2 font-heading text-sm md:text-base font-bold leading-tight truncate text-muted-foreground ">
-              {nextMatch.home_team?.flag_url && <img src={nextMatch.home_team.flag_url} alt="" className="inline h-3.5 w-5.5 align-middle" />}
-              <span className="sm:hidden lg:inline text-foreground">{homeFifaCode}</span> {t("vs")}{" "}
-              <span className="sm:hidden lg:inline text-foreground">{awayFifaCode}</span>
-              {nextMatch.away_team?.flag_url && <img src={nextMatch.away_team.flag_url} alt="" className="inline h-3.5 w-5.5 align-middle" />}
+            <div className="flex items-center sm:justify-center xl:justify-start gap-2 font-heading text-sm md:text-base font-bold leading-tight truncate text-muted-foreground">
+              {homeTeam?.flag_url && <img src={homeTeam.flag_url} alt="" className="inline h-3.5 w-5.5 align-middle" />}
+              <span className="sm:hidden lg:inline text-foreground">{homeTeam?.fifa_code ?? ""}</span> {t("vs")}{" "}
+              <span className="sm:hidden lg:inline text-foreground">{awayTeam?.fifa_code ?? ""}</span>
+              {awayTeam?.flag_url && <img src={awayTeam.flag_url} alt="" className="inline h-3.5 w-5.5 align-middle" />}
             </div>
           }
         />
